Add files via upload
This is the EDA i've done so far. I talked with the professor and he showed me how to upload the dataset, since we don't have enough storage to upload it as a pandas df. 

I haven't dealt with outliers yet, the code I have keeps returning an error, so If anyone want to take a look that'd be helpful.